+++ test/unit/parse_conditions_test.lua
@@ -14,7 +14,7 @@ g.test_parse = function()
         {'>=', 'f', {3, 3, 4}},
     }
 
-    local conditions, err = compare_conditions.parse(user_conditions)
+    local conditions, err = select_conditions.parse(user_conditions)
     t.assert(err == nil)
     t.assert_equals(conditions, {
         cond_funcs.eq('aaaa', nil),
@@ -30,7 +30,7 @@ g.test_parse_errors = function()
     -- conditions are no table
     local user_conditions = 'bbb = {12}'
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(err.err, 'Conditions should be table, got "string"')
 
     -- condition is no table
@@ -39,7 +39,7 @@ g.test_parse_errors = function()
         'bbb = {12}',
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(err.err, 'Each condition should be table, got "string" (condition 2)')
 
     -- condition len is wrong
@@ -48,7 +48,7 @@ g.test_parse_errors = function()
         {'='},
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(
         err.err,
         'Each condition should be {"<operator>", "<operand>", <value>} (condition 2)'
@@ -59,7 +59,7 @@ g.test_parse_errors = function()
         {'=', 'bb', 1, 2},
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(
         err.err,
         'Each condition should be {"<operator>", "<operand>", <value>} (condition 2)'
@@ -71,7 +71,7 @@ g.test_parse_errors = function()
         {3, 'bb', 1},
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(
         err.err,
         'condition[1] should be string, got "number" (condition 2)'
@@ -83,7 +83,7 @@ g.test_parse_errors = function()
         {'===', 'bb', 1},
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(
         err.err,
         'condition[1] "===" isn\'t a valid condition oprator, (condition 2)'
@@ -95,9 +95,81 @@ g.test_parse_errors = function()
         {'=', 3, 1},
     }
 
-    local _, err = compare_conditions.parse(user_conditions)
+    local _, err = select_conditions.parse(user_conditions)
     t.assert_str_contains(
         err.err,
         'condition[2] should be string, got "number" (condition 2)'
+    )
+end
+
+g.test_jsonpath_parse = function()
+    t.skip_if(not crud_utils.tarantool_supports_jsonpath_filters(), "Jsonpath is not supported on Tarantool < 1.10")
+
+    local user_conditions = {
+        {'==', '[\'name\']', 'Alexey'},
+        {'=', '["name"].a.b', 'Sergey'},
+        {'<', '["year"]["field_1"][\'field_2\']', 2021},
+        {'<=', '[2].a', {1, 2, 3}},
+        {'>', '[2]', 'Jackson'},
+        {'>=', '[\'year\'].a["f2"][\'f3\']', 2017},
+    }
+
+    local space_format = {
+        {name = 'name', type = 'string'},
+        {name = 'surname', type = 'any'},
+        {name = 'year', type ='unsigned'},
+    }
+
+    local conditions, err = select_conditions.parse(user_conditions, space_format)
+    t.assert(err == nil)
+    t.assert_equals(conditions, {
+        cond_funcs.eq('name', 'Alexey'),
+        cond_funcs.eq('name', 'Sergey', '[1].a.b'),
+        cond_funcs.lt('year', 2021, '[3]["field_1"][\'field_2\']'),
+        cond_funcs.le('surname', {1, 2, 3}, '[2].a'),
+        cond_funcs.gt('surname', 'Jackson'),
+        cond_funcs.ge('year', 2017, '[3].a["f2"][\'f3\']'),
+    })
+end
+
+g.test_jsonpath_parse_errors = function()
+    t.skip_if(not crud_utils.tarantool_supports_jsonpaths(), "Jsonpath is not supported on Tarantool < 2")
+    local space_format = {
+        {name = 'name', type = 'string'},
+        {name = 'surname', type = 'any'},
+        {name = 'year', type ='unsigned'},
+    }
+
+    -- bad jsonpath
+    local user_conditions = {
+        {'==', '1].a', 'Alexey'},
+    }
+
+    local _, err = select_conditions.parse(user_conditions, space_format)
+    t.assert_str_contains(
+        err.err,
+        'Invalid jsonpath format'
+    )
+
+    -- non-existen fieldno
+    local user_conditions = {
+         {'==', '[4].a.b', 88},
+    }
+
+    local _, err = select_conditions.parse(user_conditions, space_format)
+    t.assert_str_contains(
+        err.err,
+        'Space doesn\'t contains field [4]'
+    )
+
+    -- non-existen field
+    local user_conditions = {
+        {'==', '[\'bucket_id\']', 41},
+    }
+
+    local _, err = select_conditions.parse(user_conditions, space_format)
+    t.assert_str_contains(
+        err.err,
+        'Space doesn\'t contains field [\'bucket_id\']'
     )
 end