+++ test/unit/select_filters_test.lua
@@ -83,8 +83,13 @@ g.test_parse = function()
     -- age filter (early exit is possible)
     local age_filter_condition = filter_conditions[1]
     t.assert_type(age_filter_condition, 'table')
+<<<<<<< HEAD
     t.assert_equals(age_filter_condition.fieldnos, {5})
     t.assert_equals(age_filter_condition.operator, compare_conditions.operators.LT)
+=======
+    t.assert_equals(age_filter_condition.fields, {5})
+    t.assert_equals(age_filter_condition.operator, select_conditions.operators.LT)
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
     t.assert_equals(age_filter_condition.values, {40})
     t.assert_equals(age_filter_condition.types, {'number'})
     t.assert_equals(age_filter_condition.early_exit_is_possible, true)
@@ -92,8 +97,13 @@ g.test_parse = function()
     -- full_name filter
     local full_name_filter_condition = filter_conditions[2]
     t.assert_type(full_name_filter_condition, 'table')
+<<<<<<< HEAD
     t.assert_equals(full_name_filter_condition.fieldnos, {3, 4})
     t.assert_equals(full_name_filter_condition.operator, compare_conditions.operators.EQ)
+=======
+    t.assert_equals(full_name_filter_condition.fields, {3, 4})
+    t.assert_equals(full_name_filter_condition.operator, select_conditions.operators.EQ)
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
     t.assert_equals(full_name_filter_condition.values, {'Ivan', 'Ivanov'})
     t.assert_equals(full_name_filter_condition.types, {'string', 'string'})
     t.assert_equals(full_name_filter_condition.early_exit_is_possible, false)
@@ -115,8 +125,13 @@ g.test_parse = function()
     -- has_a_car filter
     local has_a_car_filter_condition = filter_conditions[3]
     t.assert_type(has_a_car_filter_condition, 'table')
+<<<<<<< HEAD
     t.assert_equals(has_a_car_filter_condition.fieldnos, {7})
     t.assert_equals(has_a_car_filter_condition.operator, compare_conditions.operators.EQ)
+=======
+    t.assert_equals(has_a_car_filter_condition.fields, {7})
+    t.assert_equals(has_a_car_filter_condition.operator, select_conditions.operators.EQ)
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
     t.assert_equals(has_a_car_filter_condition.values, {true})
     t.assert_equals(has_a_car_filter_condition.types, {'boolean'})
     t.assert_equals(has_a_car_filter_condition.early_exit_is_possible, false)
@@ -130,8 +145,13 @@ end
 g.test_one_condition_number = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {3},
             types = {'number'},
             early_exit_is_possible = true,
@@ -167,8 +187,13 @@ end
 g.test_one_condition_boolean = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {true},
             types = {'boolean'},
             early_exit_is_possible = true,
@@ -207,8 +232,13 @@ end
 g.test_one_condition_string = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {2},
             operator = compare_conditions.operators.GT,
+=======
+            fields = {2},
+            operator = select_conditions.operators.GT,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'dddddddd'},
             types = {'string'},
             early_exit_is_possible = true,
@@ -248,15 +278,25 @@ end
 g.test_two_conditions = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {4},
             types = {'number'},
             early_exit_is_possible = true,
         },
         {
+<<<<<<< HEAD
             fieldnos = {3},
             operator = compare_conditions.operators.GE,
+=======
+            fields = {3},
+            operator = select_conditions.operators.GE,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {"dddddddd"},
             types = {'string'},
             early_exit_is_possible = false,
@@ -304,25 +344,35 @@ end
 g.test_two_conditions_non_nullable = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {2, 3},
             operator = compare_conditions.operators.GE,
+=======
+            fields = {2, 3},
+            operator = select_conditions.operators.GE,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {"test", 5},
             types = {'string', 'number'},
             early_exit_is_possible = false,
             values_opts = {
                 {is_nullable = false},
                 {is_nullable = true},
-            }
+            },
         },
         {
+<<<<<<< HEAD
             fieldnos = {1},
             operator = compare_conditions.operators.LT,
+=======
+            fields = {1},
+            operator = select_conditions.operators.LT,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {3},
             types = {'number'},
             early_exit_is_possible = true,
             values_opts = {
                 {is_nullable = false},
-            }
+            },
         },
     }
 
@@ -372,15 +422,20 @@ end
 g.test_one_condition_with_nil_value = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {2, 3},
             operator = compare_conditions.operators.GE,
+=======
+            fields = {2, 3},
+            operator = select_conditions.operators.GE,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {"test"},
             types = {'string', 'number'},
             early_exit_is_possible = false,
             values_opts = {
                 {is_nullable = false},
                 {is_nullable = true},
-            }
+            },
         },
     }
 
@@ -415,8 +470,13 @@ end
 g.test_unicode_collation = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1, 2, 3, 4},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1, 2, 3, 4},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'A', 'Á', 'Ä', 6},
             types = {'string', 'string', 'string', 'number'},
             early_exit_is_possible = false,
@@ -424,7 +484,7 @@ g.test_unicode_collation = function()
                 {collation='unicode'},
                 {collation='unicode_ci'},
                 {collation='unicode_ci'},
-            }
+            },
         },
     }
 
@@ -461,8 +521,13 @@ end
 g.test_binary_and_none_collation = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1, 2, 3},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1, 2, 3},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'A', 'B', 'C'},
             types = {'string', 'string', 'string'},
             early_exit_is_possible = false,
@@ -470,7 +535,7 @@ g.test_binary_and_none_collation = function()
                 {collation='none'},
                 {collation='binary'},
                 {collation=nil},
-            }
+            },
         },
     }
 
@@ -506,15 +571,20 @@ end
 g.test_null_as_last_value_eq = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1, 2},
             operator = compare_conditions.operators.EQ,
+=======
+            fields = {1, 2},
+            operator = select_conditions.operators.EQ,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'a', box.NULL},
             types = {'string', 'string'},
             early_exit_is_possible = false,
             values_opts = {
                 nil,
                 {is_nullable = true},
-            }
+            },
         },
     }
 
@@ -547,15 +617,20 @@ end
 g.test_null_as_last_value_gt = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1, 2},
             operator = compare_conditions.operators.GT,
+=======
+            fields = {1, 2},
+            operator = select_conditions.operators.GT,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'a', box.NULL},
             types = {'string', 'string'},
             early_exit_is_possible = false,
             values_opts = {
                 nil,
                 {is_nullable = true},
-            }
+            },
         },
     }
 
@@ -594,15 +669,20 @@ end
 g.test_null_as_last_value_gt_non_nullable = function()
     local filter_conditions = {
         {
+<<<<<<< HEAD
             fieldnos = {1, 2},
             operator = compare_conditions.operators.GT,
+=======
+            fields = {1, 2},
+            operator = select_conditions.operators.GT,
+>>>>>>> f143377 (Remove field validation and combine with jsonpaths)
             values = {'a', box.NULL},
             types = {'string', 'string'},
             early_exit_is_possible = false,
             values_opts = {
                 nil,
                 {is_nullable = false},
-            }
+            },
         },
     }
 
@@ -637,5 +717,183 @@ return M]]
     t.assert_equals(filter_func({'a', 'a'}), true)       -- 'a' > box.NULL
     t.assert_equals(filter_func({'a', box.NULL}), false) -- box.NULL > box.NULL is false
 end
+
+g.test_jsonpath_fields_eq = function()
+    local filter_conditions = {
+        {
+            fields = {'[2].a.b'},
+            operator = select_conditions.operators.EQ,
+            values = {55},
+            types = {'number'},
+            early_exit_is_possible = true,
+        },
+    }
+
+    local expected_code = [[local tuple = ...
+
+local field__2__a_b = tuple["[2].a.b"]
+
+if not eq_1(field__2__a_b) then return false, true end
+
+return true, false]]
+
+    local expected_library_code = [[local M = {}
+
+function M.eq_1(field__2__a_b)
+    return (eq(field__2__a_b, 55))
+end
+
+return M]]
+
+    local filter_code = select_filters.internal.gen_filter_code(filter_conditions)
+    t.assert_equals(filter_code.code, expected_code)
+    t.assert_equals(filter_code.library, expected_library_code)
+
+    local filter_func = select_filters.internal.compile(filter_code)
+    t.assert_equals({ filter_func(box.tuple.new({3, {a = {b = 55}}, 1})) }, {true, false})
+    t.assert_equals({ filter_func(box.tuple.new({3, {a = {b = 23}}, 1})) }, {false, true})
+    t.assert_equals({ filter_func(box.tuple.new({3, {a = {c = 55}}, 1})) }, {false, true})
+    t.assert_equals({ filter_func(box.tuple.new({3, nil, 1})) }, {false, true})
+end
+
+g.test_jsonpath_fields_ge = function()
+    local filter_conditions = {
+        {
+            fields = {'[2]["field_2"]'},
+            operator = select_conditions.operators.GT,
+            values = {23},
+            types = {'number'},
+            early_exit_is_possible = true,
+        },
+    }
+
+    local expected_code = [[local tuple = ...
+
+local field__2___field_2__ = tuple["[2][\"field_2\"]"]
+
+if not cmp_1(field__2___field_2__) then return false, true end
+
+return true, false]]
+
+    local expected_library_code = [[local M = {}
+
+function M.cmp_1(field__2___field_2__)
+    if lt(field__2___field_2__, 23) then return false end
+    if not eq(field__2___field_2__, 23) then return true end
+
+    return false
+end
+
+return M]]
+
+    local filter_code = select_filters.internal.gen_filter_code(filter_conditions)
+    t.assert_equals(filter_code.code, expected_code)
+    t.assert_equals(filter_code.library, expected_library_code)
+
+    local filter_func = select_filters.internal.compile(filter_code)
+    t.assert_equals({ filter_func(box.tuple.new({3, {field_2 = 55}, 1})) }, {true, false})
+    t.assert_equals({ filter_func(box.tuple.new({3, {field_2 = 24, field_3 = 32}, nil})) }, {true, false})
+    t.assert_equals({ filter_func(box.tuple.new({{field_2 = 59}, 173, 1})) }, {false, true})
+end
+
+g.test_several_jsonpath = function()
+    local filter_conditions = {
+        {
+            fields = {'[3]["f2"][\'f3\']', '[4].f3'},
+            operator = select_conditions.operators.EQ,
+            values = {'a', 'b'},
+            types = {'string', 'string'},
+            early_exit_is_possible = true,
+        },
+    }
+
+    local expected_code = [[local tuple = ...
+
+local field__3___f2____f3__ = tuple["[3][\"f2\"]['f3']"]
+local field__4__f3 = tuple["[4].f3"]
+
+if not eq_1(field__3___f2____f3__, field__4__f3) then return false, true end
+
+return true, false]]
+
+    local expected_library_code = [[local M = {}
+
+function M.eq_1(field__3___f2____f3__, field__4__f3)
+    return (eq(field__3___f2____f3__, "a") and eq(field__4__f3, "b"))
+end
+
+return M]]
+
+    local filter_code = select_filters.internal.gen_filter_code(filter_conditions)
+    t.assert_equals(filter_code.code, expected_code)
+    t.assert_equals(filter_code.library, expected_library_code)
+
+    local filter_func = select_filters.internal.compile(filter_code)
+    t.assert_equals({ filter_func(box.tuple.new({1, 2, {f2 = {f3 = "a"}}, {f3 = "b"}})) }, {true, false})
+    t.assert_equals({ filter_func(box.tuple.new({1, 2, {f3 = "b"}}, {f2 = {f3 = "a"}})) }, {false, true})
+    t.assert_equals({ filter_func(box.tuple.new({1, 2, {f2 = {f3 = "a"}}, "b"})) }, {false, true})
+end
+
+g.test_jsonpath_two_conditions = function()
+    local filter_conditions = {
+        {
+            fields = {'[2].fld_1', '[3]["f_1"]'},
+            operator = select_conditions.operators.GE,
+            values = {"jsonpath_test", 23},
+            types = {'string', 'number'},
+            early_exit_is_possible = false,
+        },
+        {
+            fields = {'[1].field_1'},
+            operator = select_conditions.operators.LT,
+            values = {8},
+            types = {'number'},
+            early_exit_is_possible = true,
+        },
+    }
+
+    local expected_code = [[local tuple = ...
+
+local field__2__fld_1 = tuple["[2].fld_1"]
+local field__3___f_1__ = tuple["[3][\"f_1\"]"]
+local field__1__field_1 = tuple["[1].field_1"]
+
+if not cmp_1(field__2__fld_1, field__3___f_1__) then return false, false end
+if not cmp_2(field__1__field_1) then return false, true end
+
+return true, false]]
+
+    local expected_library_code = [[local M = {}
+
+function M.cmp_1(field__2__fld_1, field__3___f_1__)
+    if lt(field__2__fld_1, "jsonpath_test") then return false end
+    if not eq(field__2__fld_1, "jsonpath_test") then return true end
+
+    if lt(field__3___f_1__, 23) then return false end
+    if not eq(field__3___f_1__, 23) then return true end
+
+    return true
+end
+
+function M.cmp_2(field__1__field_1)
+    if lt(field__1__field_1, 8) then return true end
+    if not eq(field__1__field_1, 8) then return false end
+
+    return false
+end
+
+return M]]
+
+    local filter_code = select_filters.internal.gen_filter_code(filter_conditions)
+    t.assert_equals(filter_code.code, expected_code)
+    t.assert_equals(filter_code.library, expected_library_code)
+
+    local filter_func = select_filters.internal.compile(filter_code)
+    t.assert_equals({ filter_func(box.tuple.new({{field_1 = 7}, {fld_1 = "jsonpath_test"}, {f_1 = 23}})) }, {true, false})
+    t.assert_equals({ filter_func(box.tuple.new({{field_1 = 8}, {fld_1 = "jsonpath_test"}, {f_1 = 23}})) }, {false, true})
+    t.assert_equals({ filter_func(box.tuple.new({{field_1 = 5}, {f_1 = "jsonpath_test"}, {fld_1 = 23}})) }, {false, false})
+    t.assert_equals({ filter_func(box.tuple.new({{field_1 = 5, f2 = 3}, {fld_1 = "jsonpath_test"}, 23})) }, {false, false})
+end
+
 
 -- luacheck: pop