+++ test/unit/select_filters_uuid_test.lua
@@ -73,8 +73,8 @@ g.test_parse = function()
     -- uuid filter (early exit is possible)
     local uuid_filter_condition = filter_conditions[1]
     t.assert_type(uuid_filter_condition, 'table')
-    t.assert_equals(uuid_filter_condition.fieldnos, {1})
-    t.assert_equals(uuid_filter_condition.operator, compare_conditions.operators.LT)
+    t.assert_equals(uuid_filter_condition.fields, {1})
+    t.assert_equals(uuid_filter_condition.operator, select_conditions.operators.LT)
     t.assert_equals(uuid_filter_condition.values, {uuid2})
     t.assert_equals(uuid_filter_condition.types, {'uuid'})
     t.assert_equals(uuid_filter_condition.early_exit_is_possible, true)
@@ -82,8 +82,8 @@ g.test_parse = function()
     -- name filter
     local name_filter_condition = filter_conditions[2]
     t.assert_type(name_filter_condition, 'table')
-    t.assert_equals(name_filter_condition.fieldnos, {3})
-    t.assert_equals(name_filter_condition.operator, compare_conditions.operators.EQ)
+    t.assert_equals(name_filter_condition.fields, {3})
+    t.assert_equals(name_filter_condition.operator, select_conditions.operators.EQ)
     t.assert_equals(name_filter_condition.values, {'Charlie'})
     t.assert_equals(name_filter_condition.types, {'string'})
     t.assert_equals(name_filter_condition.early_exit_is_possible, false)
@@ -91,8 +91,8 @@ g.test_parse = function()
     -- has_a_car filter
     local category_id_filter_condition = filter_conditions[3]
     t.assert_type(category_id_filter_condition, 'table')
-    t.assert_equals(category_id_filter_condition.fieldnos, {4})
-    t.assert_equals(category_id_filter_condition.operator, compare_conditions.operators.EQ)
+    t.assert_equals(category_id_filter_condition.fields, {4})
+    t.assert_equals(category_id_filter_condition.operator, select_conditions.operators.EQ)
     t.assert_equals(category_id_filter_condition.values, {uuid3})
     t.assert_equals(category_id_filter_condition.types, {'uuid'})
     t.assert_equals(category_id_filter_condition.early_exit_is_possible, false)
@@ -111,8 +111,8 @@ g.test_one_condition_uuid = function()
 
     local filter_conditions = {
         {
-            fieldnos = {1},
-            operator = compare_conditions.operators.EQ,
+            fields = {1},
+            operator = select_conditions.operators.EQ,
             values = {uuid1},
             types = {'uuid'},
             early_exit_is_possible = true,
@@ -153,8 +153,8 @@ g.test_one_condition_uuid_gt = function()
 
     local filter_conditions = {
         {
-            fieldnos = {1},
-            operator = compare_conditions.operators.GT,
+            fields = {1},
+            operator = select_conditions.operators.GT,
             values = {uuid1},
             types = {'uuid'},
             early_exit_is_possible = true,
@@ -198,15 +198,15 @@ g.test_one_condition_uuid_with_nil_value = function()
 
     local filter_conditions = {
         {
-            fieldnos = {1, 3},
-            operator = compare_conditions.operators.GE,
+            fields = {1, 3},
+            operator = select_conditions.operators.GE,
             values = {uuid1},
             types = {'uuid', 'string'},
             early_exit_is_possible = false,
             values_opts = {
                 {is_nullable = false},
                 {is_nullable = true},
-            }
+            },
         },
     }
 
@@ -246,15 +246,15 @@ g.test_two_conditions_uuid = function()
 
     local filter_conditions = {
         {
-            fieldnos = {2},
-            operator = compare_conditions.operators.EQ,
+            fields = {2},
+            operator = select_conditions.operators.EQ,
             values = {'Charlie'},
             types = {'string'},
             early_exit_is_possible = true,
         },
         {
-            fieldnos = {3},
-            operator = compare_conditions.operators.GE,
+            fields = {3},
+            operator = select_conditions.operators.GE,
             values = {uuid2:str()},
             types = {'uuid'},
             early_exit_is_possible = false,