+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Option flag `force_map_call` for `select()`/`pairs()`
   to disable the `bucket_id` computation from primary key.
 * `crud.min` and `crud.max` functions to find the minimum and maximum values in the specified index.
+* Added support for jsonpath for select.
 
 ## [0.6.0] - 2021-03-29